fix ut bug
